Add files via upload
New changes for responsive design: 
O mne - title is now scaling better for smaller displays - eliminated margin around " O Mne " and also for left and right side of the paragraph for Veronika so it will be nicer on the smaller screens with less space. Additionally, added width for 400px for only logo "Jumping s Veronikou" to scale the side down for better visual appeal.